--- src/main/java/com/example/wllserver/Entity/WasherType.java
@@ -1,6 +1,0 @@
-package com.example.wllserver.Entity;
-
-public enum WasherType {
-    WASHER,
-    DRYER
-}
