+++ src/main/java/com/example/wllserver/domain/auth/register/dto/request/RegisterRequest.java
@@ -1,0 +1,9 @@
+package com.example.wllserver.domain.auth.register.dto.request;
+
+public record RegisterRequest(
+
+        Long userid,
+        String username
+
+) {
+}
