+++ src/main/java/com/example/wllserver/domain/auth/register/dto/response/LoginResponse.java
@@ -1,0 +1,13 @@
+package com.example.wllserver.domain.auth.register.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class LoginResponse {
+    private String token;
+    private String tokenType;
+}
