+++ src/main/java/com/example/wllserver/domain/auth/register/dto/response/RegisterResponse.java
@@ -1,0 +1,12 @@
+package com.example.wllserver.domain.auth.register.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class RegisterResponse {
+    private String message;
+}
