+++ src/main/java/com/example/wllserver/domain/user/domain/entity/UserEntity.java
@@ -1,0 +1,25 @@
+package com.example.wllserver.domain.user.domain.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.experimental.SuperBuilder;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "users")
+@SuperBuilder
+public class UserEntity {
+
+    @Id
+    private Long userid;
+
+    private String username;
+
+}
