+++ src/main/java/com/example/wllserver/domain/washer/domain/enums/WasherType.java
@@ -1,0 +1,6 @@
+package com.example.wllserver.domain.washer.domain.enums;
+
+public enum WasherType {
+    WASHER,
+    DRYER
+}
