+++ src/main/java/com/example/wllserver/domain/washer/dto/response/UseWasherResponse.java
@@ -1,0 +1,8 @@
+package com.example.wllserver.domain.washer.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public class UseWasherResponse {
+
+}
