--- src/main/java/com/example/wllserver/dto/ErrorResponse.java
@@ -1,16 +1,0 @@
-package com.example.wllserver.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class ErrorResponse {
-
-    private String message;
-
-    public String error;
-
-}
