--- src/main/java/com/example/wllserver/dto/LoginResponse.java
@@ -1,13 +1,0 @@
-package com.example.wllserver.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class LoginResponse {
-    private String token;
-    private String tokenType;
-}
