--- src/main/java/com/example/wllserver/dto/RoomResponse.java
@@ -1,15 +1,0 @@
-package com.example.wllserver.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class RoomResponse {
-    private Long roomid;
-
-    public RoomResponse(Long roomid) {
-        this.roomid = roomid;
-    }
-
-}
