--- src/main/java/com/example/wllserver/dto/UseWasherResponse.java
@@ -1,8 +1,0 @@
-package com.example.wllserver.dto;
-
-import lombok.Builder;
-
-@Builder
-public class UseWasherResponse {
-
-}
