--- src/main/java/com/example/wllserver/dto/register/RegisterRequest.java
@@ -1,9 +1,0 @@
-package com.example.wllserver.dto.register;
-
-public record RegisterRequest(
-
-        Long userid,
-        String username
-
-) {
-}
