--- src/main/java/com/example/wllserver/dto/register/RegisterResponse.java
@@ -1,12 +1,0 @@
-package com.example.wllserver.dto.register;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class RegisterResponse {
-    private String message;
-}
