+++ src/main/java/com/example/wllserver/global/exception/ErrorResponse.java
@@ -1,0 +1,16 @@
+package com.example.wllserver.global.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class ErrorResponse {
+
+    private String message;
+
+    public String error;
+
+}
