--- src/main/java/com/example/wllserver/jwt/JwtFilter.java
@@ -1,57 +1,0 @@
-package com.example.wllserver.jwt;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.List;
-
-@Component
-public class JwtFilter extends OncePerRequestFilter {
-
-    @Autowired
-    private JwtUtil jwtUtil;
-
-
-    private final List<String> permitAllPaths = Arrays.asList(
-            "/register", "/login", "/swagger-ui/**", "/v3/api-docs/**",
-            "/test/permitAll", "/wash/**"
-    );
-
-    private boolean isAuthenticatedPath(String path) {
-        return permitAllPaths.stream().anyMatch(path::startsWith);
-    }
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws ServletException, IOException {
-
-        String path = request.getRequestURI();
-        if (isAuthenticatedPath(path)) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-
-        String authorizationHeader = request.getHeader("Authorization");
-
-        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
-            String token = authorizationHeader.substring(7);
-
-            jwtUtil.isTokenValid(token);
-
-            Authentication auth = jwtUtil.getAuthenticationByToken(token);
-            SecurityContextHolder.getContext().setAuthentication(auth);
-        }
-
-        filterChain.doFilter(request, response);
-    }
-
-}
