--- src/main/java/com/example/wllserver/jwt/JwtUtil.java
@@ -1,60 +1,0 @@
-package com.example.wllserver.jwt;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.stereotype.Component;
-
-import java.util.Collections;
-import java.util.Date;
-
-@Component
-public class JwtUtil {
-
-    @Value("${jwt.secret}")
-    private String secret;
-
-    public String generateToken(String id) {
-        return Jwts.builder()
-                .setSubject(id)
-                .setIssuedAt(new Date())
-                .signWith(SignatureAlgorithm.HS512, secret)
-                .compact();
-    }
-
-    public String extractUsername(String token) {
-        return extractAllClaims(token).getSubject();
-    }
-
-    private Claims extractAllClaims(String token) {
-        return Jwts.parserBuilder()
-                .setSigningKey(secret)
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-
-    private boolean isTokenExpired(String token) {
-        Date expiration = extractAllClaims(token).getExpiration();
-        return expiration != null && expiration.before(new Date());
-    }
-
-    public boolean isTokenValid(String token) {
-        return !isTokenExpired(token);
-    }
-
-    public Authentication getAuthenticationByToken(String token) {
-        Claims claims = extractAllClaims(token);
-
-        User principal = new User(claims.getSubject(), "",
-                Collections.singleton(new SimpleGrantedAuthority("user")));
-
-        return new UsernamePasswordAuthenticationToken(principal, token,
-                Collections.singleton(new SimpleGrantedAuthority("user")));
-    }
-}
