--- src/main/java/com/example/wllserver/repository/RoomRepository.java
@@ -1,7 +1,0 @@
-package com.example.wllserver.repository;
-
-import com.example.wllserver.Entity.RoomEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoomRepository extends JpaRepository<RoomEntity, Long> {
-}
