--- src/main/java/com/example/wllserver/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.wllserver.repository;
-
-import com.example.wllserver.Entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    Optional<UserEntity> findByUserid(Long userid);
-}
